+++ Entities/MOSParticle.cpp
@@ -152,6 +152,14 @@ namespace RTE {
 
 /////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
+	void MOSParticle::Update() {
+		MOSprite::Update();
+
+		if (m_pScreenEffect) { SetPostScreenEffectToDraw(); }
+	}
+
+/////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
 	void MOSParticle::Draw(BITMAP *targetBitmap, const Vector &targetPos, DrawMode mode, bool onlyPhysical) const {
 		RTEAssert(!m_aSprite.empty(), "No sprite bitmaps loaded to draw " + GetPresetName());
 		RTEAssert(m_Frame >= 0 && m_Frame < m_FrameCount, "Frame is out of bounds for " + GetPresetName());
@@ -262,10 +270,6 @@ namespace RTE {
 			g_ThreadMan.GetSimRenderQueue().push_back(renderFunc);
 		} else {
 			renderFunc(1.0F);
-		}
-
-		if (m_pScreenEffect && mode == g_DrawColor && !onlyPhysical) { 
-			SetPostScreenEffectToDraw(); 
 		}
 	}