+++ Entities/Scene.cpp
@@ -18,6 +18,7 @@
 #include "FrameMan.h"
 #include "ConsoleMan.h"
 #include "SettingsMan.h"
+#include "ThreadMan.h"
 #include "MetaMan.h"
 #include "ContentFile.h"
 #include "SLTerrain.h"
@@ -43,6 +44,8 @@
 #include "HDFirearm.h"
 #include "Magazine.h"
 #include "ThrownDevice.h"
+
+#include "tracy/Tracy.hpp"
 
 namespace RTE {
 
@@ -2967,6 +2970,8 @@ void Scene::BlockUntilAllPathingRequestsComplete() {
 
 void Scene::UpdatePathFinding()
 {
+    ZoneScoped;
+    
     constexpr int nodeUpdatesPerCall = 100;
     constexpr int maxUnupdatedMaterialAreas = 1000;
 
@@ -3098,6 +3103,8 @@ void Scene::Unlock()
 //                  before drawing.
 
 void Scene::UpdateSim() {
+    ZoneScoped;
+
     m_PathfindingUpdated = false;
 
 	if (g_SettingsMan.BlipOnRevealUnseen())
@@ -3116,6 +3123,9 @@ void Scene::UpdateSim() {
 	}
 
     if (m_NavigatableAreasUpToDate == false) {
+        // Need to block until all current pathfinding requests are finished. Ugh, if only we had a better way (interrupt/cancel a path request to start a new one?)
+        g_ThreadMan.GetBackgroundThreadPool().wait_for_tasks();
+
         m_NavigatableAreasUpToDate = true;
         for (int team = Activity::Teams::NoTeam; team < Activity::Teams::MaxTeamCount; ++team) {
             PathFinder& pathFinder = *GetPathFinder(static_cast<Activity::Teams>(team));