+++ Main.cpp
@@ -47,6 +47,7 @@
 #include "CameraMan.h"
 #include "ActivityMan.h"
 #include "PrimitiveMan.h"
+#include "ThreadMan.h"
 
 #include "tracy/Tracy.hpp"
 
@@ -62,6 +63,7 @@ namespace RTE {
 	/// Initializes all the essential managers.
 	/// </summary>
 	void InitializeManagers() {
+		ThreadMan::Construct();
 		TimerMan::Construct();
 		PresetMan::Construct();
 		SettingsMan::Construct();
@@ -86,6 +88,7 @@ namespace RTE {
 		LoadingScreen::Construct();
 		ThreadMan::Construct();
 
+		g_ThreadMan.Initialize();
 		g_SettingsMan.Initialize();
 		g_WindowMan.Initialize();