+++ Managers/SettingsMan.h
@@ -118,12 +118,6 @@ namespace RTE {
 		bool GetAnyExperimentalSettingsEnabled() const { return false; }
 
 		/// <summary>
-		/// Returns whether or not multithreaded Lua is enabled.
-		/// </summary>
-		/// <returns>Whether or not multithreaded Lua is enabled.</returns>
-		bool GetEnableMultithreadedLua() const { return m_EnableMultithreadedLua; }
-
-		/// <summary>
 		/// Gets the AI update interval.
 		/// </summary>
 		/// <returns>How often Actor's AI is updated, in simulation updates.</returns>
@@ -541,7 +535,6 @@ namespace RTE {
 		bool m_AllowSavingToBase; //!< Whether editors will allow to select Base.rte as a module to save in.
 		bool m_ShowMetaScenes; //!< Show MetaScenes in editors and activities.
 
-		int m_EnableMultithreadedLua; //!< Whether or not to enable multithreaded Lua scripts.
 		bool m_DisableLuaJIT; //!< Whether to disable LuaJIT or not. Disabling will skip loading the JIT library entirely as just setting 'jit.off()' seems to have no visible effect.
 		int m_RecommendedMOIDCount; //!< Recommended max MOID's before removing actors from scenes.
 		int m_SceneBackgroundAutoScaleMode; //!< Scene background layer auto-scaling mode. 0 for off, 1 for fit screen dimensions and 2 for always upscaled to x2.